requirements.txt
langchain @ git+https://github.com/myscale/langchain.git@preview#egg=langchain&subdirectory=libs/langchain
langchain-experimental @ git+https://github.com/myscale/langchain.git@preview#egg=langchain-experimental&subdirectory=libs/experimental
InstructorEmbedding
pandas
sentence_transformers
streamlit==1.25
altair==4.2.2
clickhouse-connect
openai
lark
tiktoken
sql-formatter
clickhouse-sqlalchemy